+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Window start up location and size are now stored so that when you next launch ProtobufDecoder it will restore to the same size as before.
 - The help text in the "Load hex stream" dialog now shows the supported formats.
+- ProtobufDecoder now targets .Net 6.0
 
 ### Fixes